start to change handler and platform back to api model

+++ cmd/zymsrvd/handlers/chamberhandler.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"net/url"
 
 	"github.com/benjaminbartels/zymurgauge/internal"
 	"github.com/benjaminbartels/zymurgauge/internal/database"
@@ -29,37 +30,59 @@ func NewChamberHandler(repo *database.ChamberRepo, pubSub *pubsub.PubSub, logger
 	}
 }
 
-// GetAll handles a GET request for all Chambers
-func (h *ChamberHandler) GetAll(ctx context.Context, w http.ResponseWriter, r *http.Request,
-	p map[string]string) error {
-	if chambers, err := h.repo.GetAll(); err != nil {
-		return err
-	} else {
-		web.Respond(ctx, w, chambers, http.StatusOK)
+func (h *ChamberHandler) Handle(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+	switch r.Method {
+	case web.GET:
+		return h.handleGet(ctx, w, r)
+	case web.POST:
+		return h.handlePost(ctx, w, r)
+	case web.DELETE:
+		return h.handleDelete(ctx, w, r)
+	default:
+		return web.ErrMethodNotAllowed
 	}
-	return nil
 }
 
-// GetOne handles a GET request for a specific Chamber whose mac address matched the provided mac
-func (h *ChamberHandler) GetOne(ctx context.Context, w http.ResponseWriter, r *http.Request,
-	p map[string]string) error {
-	mac := p["mac"]
+func (h *ChamberHandler) handleGet(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+	var head string
+	head, r.URL.Path = web.ShiftPath(r.URL.Path)
+	if head == "" {
+		return h.handleGetAll(ctx, w)
+	} else {
+		mac, err := url.QueryUnescape(head)
+		if err != nil {
+			return web.ErrBadRequest
+		}
+		head, r.URL.Path = web.ShiftPath(r.URL.Path)
+		if head == "events" {
+			return h.handleGetEvents(ctx, w, mac)
+		} else if head == "" {
+			return h.handleGetOne(ctx, w, mac)
+		} else {
+			return web.ErrBadRequest
+		}
+	}
+}
+
+func (h *ChamberHandler) handleGetOne(ctx context.Context, w http.ResponseWriter, mac string) error {
 	if chamber, err := h.repo.Get(mac); err != nil {
 		return err
 	} else if chamber == nil {
 		return web.ErrNotFound
 	} else {
-		web.Respond(ctx, w, chamber, http.StatusOK)
+		return web.Respond(ctx, w, chamber, http.StatusOK)
 	}
-	return nil
 }
 
-// GetEvents handles a GET request to listen for web events for a specific Chamber
-// whose mac address matched the provided mac
-func (h *ChamberHandler) GetEvents(ctx context.Context, w http.ResponseWriter, r *http.Request,
-	p map[string]string) error {
-	mac := p["mac"]
+func (h *ChamberHandler) handleGetAll(ctx context.Context, w http.ResponseWriter) error {
+	if chambers, err := h.repo.GetAll(); err != nil {
+		return err
+	} else {
+		return web.Respond(ctx, w, chambers, http.StatusOK)
+	}
+}
 
+func (h *ChamberHandler) handleGetEvents(ctx context.Context, w http.ResponseWriter, mac string) error {
 	f, ok := w.(http.Flusher)
 	if !ok {
 		return web.ErrInternal
@@ -81,41 +104,47 @@ func (h *ChamberHandler) GetEvents(ctx context.Context, w http.ResponseWriter, r
 			break
 		}
 		msg := fmt.Sprintf("data: %s\n", c)
+		fmt.Fprint(w, msg)
 		h.logger.Printf("Sending: %s\n", msg)
 		f.Flush()
 	}
-
 	return nil
 }
 
-// Post handles the POST request to create or update a Chamber
-func (h *ChamberHandler) Post(ctx context.Context, w http.ResponseWriter, r *http.Request, p map[string]string) error {
+func (h *ChamberHandler) handlePost(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
 	chamber, err := parseChamber(r)
 	if err != nil {
 		return err
 	}
+	if err = h.repo.Save(&chamber); err != nil {
+		return err
+	}
+	b, err := json.Marshal(chamber)
+	if err != nil {
+		return err
+	}
+	h.pubSub.Send(chamber.MacAddress, b)
+	if _, err = w.Write(b); err != nil {
+		return err
+	}
+	return nil
+}
 
-	if err := h.repo.Save(&chamber); err != nil {
+func (h *ChamberHandler) handleDelete(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+	if r.URL.Path == "" {
+		return web.ErrBadRequest
+	}
+
+	if mac, err := url.QueryUnescape(r.URL.Path); err != nil {
 		return err
 	} else {
-		web.Respond(ctx, w, chamber, http.StatusOK)
+		if err := h.repo.Delete(mac); err != nil {
+			return err
+		}
 	}
 	return nil
 }
 
-// Delete handles the DELETE request to delete a Chamber
-func (h *ChamberHandler) Delete(ctx context.Context, w http.ResponseWriter, r *http.Request, p map[string]string) error {
-	mac := p["mac"]
-
-	if err := h.repo.Delete(mac); err != nil {
-		return err
-	}
-
-	web.Respond(ctx, w, nil, http.StatusOK)
-	return nil
-}
-
-// parseChamber decodes the specified Chamber into JSON
 func parseChamber(r *http.Request) (internal.Chamber, error) {
 	var chamber internal.Chamber
 	err := json.NewDecoder(r.Body).Decode(&chamber)
